+++ source/XBee.Core/Api/Features/PacketListenning/NodeDiscoveryListener.cs
@@ -4,7 +4,7 @@ namespace Gadgeteer.Modules.GHIElectronics.Api
 {
     public class NodeDiscoveryListener : MultiplePacketListener
     {
-        public NodeDiscoveryListener(int maxNodes = int.MaxValue)
+        public NodeDiscoveryListener(byte maxNodes = byte.MaxValue)
             : base(maxNodes, null)
         {
             Validator = new AtCommandValidator(AtCmd.NodeDiscover);
