+++ source/XBee.Core/Util/IInputStream.cs
@@ -4,8 +4,8 @@ namespace Gadgeteer.Modules.GHIElectronics.Util
 {
     public interface IInputStream : IDisposable
     {
-        int Read();
-        int Read(string message);
-        int[] Read(int count);
+        byte Read();
+        byte Read(string message);
+        byte[] Read(int count);
     }
 }
